Fixed premature return in FilterBuilder.includePackage(String...)
FilterBuilder.includePackage(String...) was returning after adding the first package
passed in.  As a result, if the method was called with more than one package, only
the first would be added.  All others would be ignored.  This fix removes the errant
return statment in the for loop to ensure that all passed in packages will be added.